+++ assets/js/blocks/cart-checkout/checkout/frontend.js
@@ -7,7 +7,7 @@ import { withRestApiHydration } from '@woocommerce/block-hocs';
  * Internal dependencies
  */
 import Block from './block.js';
-import renderFrontend from '../../utils/render-frontend.js';
+import renderFrontend from '../../../utils/render-frontend.js';
 
 const getProps = () => {
 	return {
